+++ config/routes.rb
@@ -251,7 +251,7 @@ Rails.application.routes.draw do
   end
 
   # Project groups
-  resources :groups, only: %i[index], controller: :project_groups, as: :project_groups
+  resources :groups, only: %i[index show], controller: :project_groups, as: :project_groups
 
   # Service support projects
   constraints(id: VALID_UUID_REGEX) do