+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - show an indicator when a project is part of a group
 - a new index page listing all grouped projects
 - a list of the project groups is shown on the index page
+- individual project groups can now be viewed
 
 ## [Release-81][release-81]